Changed the controller to not have the authorized check anymore, but every route that needs it is secured with the authorisation check.

+++ api/src/main/java/com/cloudmen/cloudguard/dto/groups/GroupOverviewResponse.java
@@ -4,11 +4,11 @@ import com.cloudmen.cloudguard.dto.password.SecurityScoreBreakdownDto;
 import com.cloudmen.cloudguard.dto.preferences.SectionWarningsDto;
 
 public record GroupOverviewResponse(
-        long totalGroups,
-        long groupsWithExternal,
-        long highRiskGroups,
-        long mediumRiskGroups,
-        long lowRiskGroups,
+        int totalGroups,
+        int groupsWithExternal,
+        int highRiskGroups,
+        int mediumRiskGroups,
+        int lowRiskGroups,
         int securityScore,
         SecurityScoreBreakdownDto securityScoreBreakdown,
         SectionWarningsDto warnings
